Fix drizzle initialization for react-native
React-native does not support async Drizzle initialization via a promise and window.addEventListener, we therefore check to make sure that we are not running in react-native before attempting this. If we are running in react-native we just dispatch the initialize action immediately.

It's probably possible to hook some sort of other event in react-native to allow async loading of Drizzle which would in theory allow for injection of Web3 but I don't know how much sense injection makes in the context of a mobile device.

It's worth noting that currently this code does not mess with Node.JS support in any way but this might also benefit from such a fix.